Renamed InfiniteCmd.java into DefaultCmd for a clearer use and added a way to disable the command.
