+++ TeamCode/src/org/firstinspires/ftc/teamcode/tests/DriveSubsystem.java
@@ -2,8 +2,7 @@ package org.firstinspires.ftc.teamcode.tests;
 
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import org.firstinspires.ftc.teamcode.microcmd.cmd.Cmd;
-import org.firstinspires.ftc.teamcode.microcmd.cmd.InfiniteCmd;
-import org.firstinspires.ftc.teamcode.microcmd.cmd.InstantCmd;
+import org.firstinspires.ftc.teamcode.microcmd.cmd.DefaultCmd;
 import org.firstinspires.ftc.teamcode.microcmd.subsystem.Subsystem;
 
 import java.util.function.DoubleSupplier;
@@ -22,7 +21,7 @@ public class DriveSubsystem extends Subsystem {
     }
 
     public Cmd drive(DoubleSupplier x, DoubleSupplier y, DoubleSupplier turn) {
-        return new InfiniteCmd(() -> {
+        return new DefaultCmd(() -> {
             frontLeft.setPower(-y.getAsDouble() + x.getAsDouble() + turn.getAsDouble());
             frontRight.setPower(-y.getAsDouble() - x.getAsDouble() - turn.getAsDouble());
             backLeft.setPower(-y.getAsDouble() - x.getAsDouble() + turn.getAsDouble());
@@ -30,6 +29,6 @@ public class DriveSubsystem extends Subsystem {
             System.out.println(x.getAsDouble());
             System.out.println(y.getAsDouble());
             System.out.println(turn.getAsDouble());
-        });
+        }, () -> false);
     }
 }
